feat: ✨Refactor task addition; implement addTask function and update DashboardScreen
