+++ lib/features/dashboard/presentation/widgets/dashboard_mixin.dart
@@ -1,102 +1,6 @@
-// TODO: Remove after fixing
-// ignore_for_file: unused_import
-
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:uuid/uuid.dart';
-
-import 'package:to_do_now/features/task/task.dart';
-import 'package:to_do_now/routes/routes.dart';
-
-import '../../../../core/core.dart';
 
 mixin DashboardMixin {
-  final formKey = GlobalKey<FormBuilderState>();
-  final TextEditingController titleController = TextEditingController();
-  final TextEditingController descriptionController = TextEditingController();
-
-  void createTask(BuildContext context) {
-    DateTime? selectedDateTIme = DateTime.now();
-    showModalBottomSheet(
-      context: context,
-      backgroundColor: AppColors.background,
-      builder: (context) {
-        return Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0, vertical: 32.0),
-          child: FormBuilder(
-            key: formKey,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              spacing: 16.0,
-              children: [
-                Text("Add Task", style: AppTypography.headline2()),
-                CustomTextFormField(
-                  name: 'title',
-                  controller: titleController,
-                  fillColor: Colors.transparent,
-                  hintText: "Title",
-                  textCapitalization: TextCapitalization.sentences,
-                  textInputAction: TextInputAction.next,
-                ),
-                CustomTextFormField(
-                  name: 'description',
-                  controller: descriptionController,
-                  hintText: "Description",
-                  fillColor: Colors.transparent,
-                  textCapitalization: TextCapitalization.sentences,
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      spacing: 24.0,
-                      children: [
-                        IconButton(
-                          icon: Icon(Icons.timer_outlined),
-                          onPressed: () async {
-                            DateTime? pickedDateTime =
-                                await _pickDateTime(context);
-                            selectedDateTIme = pickedDateTime ?? DateTime.now();
-                          },
-                        ),
-                        Icon(Icons.sell_outlined),
-                        Icon(Icons.flag_outlined),
-                      ],
-                    ),
-                    IconButton(
-                      onPressed: () {
-                        var uuid = Uuid();
-                        // TODO: fix
-                        // final task = TaskModel(
-                        //   id: uuid.v4(),
-                        //   title: titleController.text,
-                        //   description: descriptionController.text,
-                        //   dueDateTime: selectedDateTIme,
-                        // );
-
-                        // log.d("(Dashboard) New Task: $task");
-
-                        // context.read<TaskCubit>().addTask(task);
-
-                        titleController.clear();
-                        descriptionController.clear();
-
-                        context.pop();
-                      },
-                      icon: Icon(Icons.send_outlined),
-                    ),
-                  ],
-                )
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-
   Future<DateTime?> _pickDateTime(BuildContext context) async {
     DateTime? pickedDate = await showDatePicker(
       context: context,
